requirements.txt
Django==5.0.2
djangorestframework==3.14.0
dj-rest-auth==5.0.2
django-allauth==0.61.1
drf-spectacular==0.27.1
pytest==8.0.2
pytest-django==4.8.0
python-dotenv==1.0.1
Pillow==10.2.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-environ==0.11.2
gunicorn==21.2.0 
Pillow
drf-yasg

